+++ CHANGES.rst
@@ -9,6 +9,7 @@ Change History
 - Issue #7: Fixed non-recursive links.
 - Issue #8: Fixed edge cases in gitignore patterns.
 - Issue #9: Fixed minor usage documentation.
+- Fixed recursion detection.
 
 
 0.3.3 (2014-11-21)